+++ ProyectoPolideportivo/src/Mejoras
@@ -1,0 +1,2 @@
+Añadir singletone a file manager
+Trozear el main
