+++ ProyectoPolideportivo/src/main/java/AppOficina.java
@@ -1,0 +1,6 @@
+import java.util.Scanner;
+
+public class AppOficina {
+    public static void appOficina(Scanner sc){}    //}
+
+}
